requirements.txt
flask==2.3.3
python-dotenv==1.0.0
ccxt==4.1.13
pyyaml==6.0.1